refactor: Python语法现代化(1121处UP规则) + 启用UP/RUF100/SIM117门禁
UP规则自动修复(ruff --select UP --fix, 1121处):
- UP045(954): Optional[X] → X | None (PEP 604 联合类型)
- UP006(46): List/Dict/Tuple → list/dict/tuple (PEP 585 泛型)
- UP017(48): datetime.timezone.utc → datetime.UTC
- UP041(7): TimeoutError → OSError 别名
- UP035(15): 废弃typing导入清理(级联F401修复107处)
- UP042(2): str,Enum → StrEnum (models.py)

手动修复:
- core/locks.py: AsyncContextManager → contextlib.AbstractAsyncContextManager
- core/auth.py: 清除残留Optional导入
- test_finance_routes.py: 修复F811 timezone重定义(UP017级联)
- main.py: _mock_store重导出改用__all__声明(防RUF100+F401联合误删)

ruff.toml门禁升级:
- select新增 UP/RUF100/SIM117
- ruff check . = All checks passed
- 498项测试无回归(296+57+67+78)

+++ zhuxiang-jiu/backend/test_recycle_routes.py
@@ -323,7 +323,7 @@ class TestExchange:
                f"expected diff<0/points>0, got {result['priceDiff']}/{result['pointsConverted']}")
 
         # test 22: 差额转积分正确(250元 × 10 = 2500竹叶)
-        expected_points = int(250 * 10)
+        expected_points = 250 * 10
         record("test_22_points_conversion",
                result["pointsConverted"] == expected_points,
                f"expected {expected_points}, got {result['pointsConverted']}")
